crudauth/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.11.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
